PowerSystems.HydroGen Formulations

Valid DeviceModels for subtypes of HydroGen include the following:

Valid DeviceModel	Device Type	Formulation

DeviceModel(HydroDispatch, FixedOutput)	HydroDispatch	FixedOutput

DeviceModel(HydroEnergyReservoir, FixedOutput)	HydroEnergyReservoir	FixedOutput

DeviceModel(HydroPumpedStorage, FixedOutput)	HydroPumpedStorage	FixedOutput

DeviceModel(HydroDispatch, HydroDispatchRunOfRiver)	HydroDispatch	HydroDispatchRunOfRiver

DeviceModel(HydroEnergyReservoir, HydroDispatchRunOfRiver)	HydroEnergyReservoir	HydroDispatchRunOfRiver

DeviceModel(HydroPumpedStorage, HydroDispatchRunOfRiver)	HydroPumpedStorage	HydroDispatchRunOfRiver

DeviceModel(HydroDispatch, HydroDispatchPumpedStorage)	HydroDispatch	HydroDispatchPumpedStorage

DeviceModel(HydroEnergyReservoir, HydroDispatchPumpedStorage)	HydroEnergyReservoir	HydroDispatchPumpedStorage

DeviceModel(HydroPumpedStorage, HydroDispatchPumpedStorage)	HydroPumpedStorage	HydroDispatchPumpedStorage

DeviceModel(HydroDispatch, HydroDispatchReservoirBudget)	HydroDispatch	HydroDispatchReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroDispatchReservoirBudget)	HydroEnergyReservoir	HydroDispatchReservoirBudget

DeviceModel(HydroPumpedStorage, HydroDispatchReservoirBudget)	HydroPumpedStorage	HydroDispatchReservoirBudget

DeviceModel(HydroDispatch, HydroDispatchReservoirStorage)	HydroDispatch	HydroDispatchReservoirStorage

DeviceModel(HydroEnergyReservoir, HydroDispatchReservoirStorage)	HydroEnergyReservoir	HydroDispatchReservoirStorage

DeviceModel(HydroPumpedStorage, HydroDispatchReservoirStorage)	HydroPumpedStorage	HydroDispatchReservoirStorage

DeviceModel(HydroDispatch, HydroCommitmentReservoirBudget)	HydroDispatch	HydroCommitmentReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroCommitmentReservoirBudget)	HydroEnergyReservoir	HydroCommitmentReservoirBudget

DeviceModel(HydroPumpedStorage, HydroCommitmentReservoirBudget)	HydroPumpedStorage	HydroCommitmentReservoirBudget

DeviceModel(HydroEnergyReservoir, HydroCommitmentReservoirStorage)	HydroEnergyReservoir	HydroCommitmentReservoirStorage

DeviceModel(HydroDispatch, HydroCommitmentRunOfRiver)	HydroDispatch	HydroCommitmentRunOfRiver

DeviceModel(HydroEnergyReservoir, HydroCommitmentRunOfRiver)	HydroEnergyReservoir	HydroCommitmentRunOfRiver

DeviceModel(HydroPumpedStorage, HydroCommitmentRunOfRiver)	HydroPumpedStorage	HydroCommitmentRunOfRiver

HydroDispatchRunOfRiver

PowerSimulations.HydroDispatchRunOfRiver — Type

Formulation type to add injection variables constrained by a maximum injection time series for HydroGen

HydroDispatchRunOfRiver()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:53.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$.

Constraints:

\[\begin{aligned} & Pg^\text{min} \le Pg_t \le ActivePowerTimeSeriesParameter_t \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \end{aligned}\]

HydroDispatchPumpedStorage

PowerSimulations.HydroDispatchPumpedStorage — Type

Formulation type to constrain energy production from pumped storage with a representation of the energy storage capacity of upper and lower reservoirs and water inflow time series of upper reservoir and outflow time series of lower reservoir for HydroPumpedStorage

HydroDispatchPumpedStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:68.

Variables:

Static Parameters:

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.InflowTimeSeriesParameter	inflow

PowerSimulations.OutflowTimeSeriesParameter	outflow

Objective:

Constraints:

HydroDispatchReservoirBudget

PowerSimulations.HydroDispatchReservoirBudget — Type

Formulation type to add injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroDispatchReservoirBudget()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:58.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

Auxillary Variables:

EnergyOutput - TODO

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

Time Series Parameters:

Parameter	Default Time Series Name

EnergyBudgetTimeSeriesParameter	hydro_budget

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$. TODO: add slack terms

Constraints:

TODOl: add slack terms

\[\begin{aligned} & Pg^\text{min} \le Pg_t \le Pg^\text{max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \\ & \sum_{t = 1}^N(Pg_t) \cdot \Delta T \le \sum_{t = 1}^N(EnergyBudgetTimeSeriesParameter_t) \cdot \Delta T \end{aligned}\]

HydroDispatchReservoirStorage

PowerSimulations.HydroDispatchReservoirStorage — Type

Formulation type to constrain hydropower production with a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroDispatchReservoirStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:63.

Variables:

ActivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_active_power(device)

ReactivePowerVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_reactive_power(device)

EnergyVariable:

Bounds: [0.0, ]

Default initial value: PowerSystems.get_initial_storage(device)

WaterSpillageVariable:

Bounds: [0.0, ]

Default initial value: 0.0

EnergyShortageVariable:

Bounds: [0.0, ]

Default initial value: 0.0

EnergySurplusVariable:

Bounds: [0.0, ]

Default initial value: 0.0

Auxillary Variables:

EnergyOutput

Static Parameters:

$Pg^\text{min}$ = PowerSystems.get_active_power_limits(device).min

$Pg^\text{max}$ = PowerSystems.get_active_power_limits(device).max

$Qg^\text{min}$ = PowerSystems.get_reactive_power_limits(device).min

$Qg^\text{max}$ = PowerSystems.get_reactive_power_limits(device).max

$Eg^\text{max}$ = PowerSystems.get_storage_capacity(device)

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.InflowTimeSeriesParameter	inflow

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$. TODO: add slack terms

Constraints:

TODO: Add slack terms

\[\begin{aligned} & E_{t+1} = E_t + (InflowTimeSeriesParameter_t - S_t - Pg_t) \cdot \Delta T \\ & Pg^\text{min} \le Pg_t \le Pg^\text{max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \end{aligned}\]

Future releases will also implement a requirement of the energy at the last time point $N$:

\[\begin{aligned} & E_N \ge E^\text{requirement} \end{aligned}\]

HydroCommitmentReservoirBudget

PowerSimulations.HydroCommitmentReservoirBudget — Type

Formulation type to add commitment and injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroCommitmentReservoirBudget()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:78.

Variables:

Static Parameters:

Time Series Parameters:

Parameter	Default Time Series Name

EnergyBudgetTimeSeriesParameter	hydro_budget

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$. TODO: add slack and UC terms

Constraints:

Similar to the dispatch formulation, but considering a binary variable $u_t \in \{0, 1\}$ with semi continuous constraints for both active and reactive power:

\[\begin{aligned} & \sum_{t = 1}^N P_t \cdot \Delta T \le E^\text{budget} \\ & P_t - u_t P^\text{max} \le 0 \\ & P_t - u_t P^\text{min} \ge 0 \\ & Q_t - u_t Q^\text{max} \le 0 \\ & Q_t - u_t Q^\text{min} \ge 0 \end{aligned}\]

HydroCommitmentReservoirStorage

PowerSimulations.HydroCommitmentReservoirStorage — Type

Formulation type to constrain hydropower production with unit commitment variables and a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroCommitmentReservoirStorage()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:83.

Variables:

Static Parameters:

Time Series Parameters:

Parameter	Default Time Series Name

PowerSimulations.InflowTimeSeriesParameter	inflow

EnergyTargetTimeSeriesParameter	storage_target

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$. TODO: add slack and UC terms

Constraints:

TODO: Add slack and UC terms

\[\begin{aligned} & E_{t+1} = E_t + (InflowTimeSeriesParameter_t - S_t - Pg_t) \cdot \Delta T \\ & Pg^\text{min} \le Pg_t \le Pg^\text{max} \\ & Qg^\text{min} \le Qg_t \le Qg^\text{max} \end{aligned}\]

Future releases will also implement a requirement of the energy at the last time point $N$:

\[\begin{aligned} & E_N \ge E^\text{requirement} \end{aligned}\]

Similar to the dispatch formulation, but considering a binary variable $u_t \in \{0, 1\}$ with semi continuous constraints for both active and reactive power:

\[\begin{aligned} & E_{t+1} = E_t + (I_t - S_t - P_t)\Delta T \\ & P_t - u_t P^\text{max} \le 0 \\ & P_t - u_t P^\text{min} \ge 0 \\ & Q_t - u_t Q^\text{max} \le 0 \\ & Q_t - u_t Q^\text{min} \ge 0 \end{aligned}\]

HydroCommitmentRunOfRiver

PowerSimulations.HydroCommitmentRunOfRiver — Type

Formulation type to add commitment and injection variables constrained by a maximum injection time series for HydroGen

HydroCommitmentRunOfRiver()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:73.

Variables:

Static Parameters:

Time Series Parameters:

Parameter	Default Time Series Name

ActivePowerTimeSeriesParameter	max_active_power

ReactivePowerTimeSeriesParameter	max_active_power

Objective:

Creates an objective function term based on the VariableCost Options where the quantity term is defined as $Pg_t$. TODO: add UC terms

Constraints:

Similar to the dispatch formulation, but considering a binary variable $u_t \in \{0, 1\}$ with semi continuous constraints for both active and reactive power:

\[\begin{aligned} & P_t \le \eta_t P^\text{max}\\ & P_t - u_t P^\text{max} \le 0 \\ & P_t - u_t P^\text{min} \ge 0 \\ & Q_t - u_t Q^\text{max} \le 0 \\ & Q_t - u_t Q^\text{min} \ge 0 \end{aligned}\]