PowerSystems.RenewableGen Formulations

Valid DeviceModels for subtypes of RenewableGen include the following:

Valid DeviceModels	Formulation

DeviceModel(RenewableDispatch, FixedOutput)	FixedOutput

DeviceModel(RenewableFix, FixedOutput)	FixedOutput

DeviceModel(RenewableDispatch, RenewableConstantPowerFactor)	RenewableConstantPowerFactor

DeviceModel(RenewableFix, RenewableConstantPowerFactor)	RenewableConstantPowerFactor

DeviceModel(RenewableDispatch, RenewableFullDispatch)	RenewableFullDispatch

DeviceModel(RenewableFix, RenewableFullDispatch)	RenewableFullDispatch

RenewableFullDispatch

PowerSimulations.RenewableFullDispatch — Type

Formulation type to add an injection variables bounded by max_active_power time series for RenewableGen

RenewableFullDispatch()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:69.

Variables:

$P_t \geq 0$: active power injected by renewable generator at time $t$ (MW)

$Q_t \geq 0$: reactive power injected by renewable generator at time $t$ (MW)

Parameters

$P^\text{max}_t$: maximum active power availability for generator at time $t$ (MW) - defined by max_active_power time series

Objective

Creates an objective function term based on the VariableCost Options using $G_t = P^\text{max}_t - P_t$

Constraints

\[\begin{aligned} & P^\text{min} \le P_t \le P^\text{max}_t \\ & Q^\text{min} \le Q_t \le Q^\text{max} \end{aligned}\]

RenewableConstantPowerFactor

PowerSimulations.RenewableConstantPowerFactor — Type

Formulation type to add real and reactive injection variables with constant power factor bounded by max_active_power time series for RenewableGen

RenewableConstantPowerFactor()

defined at /home/runner/work/PowerSimulations.jl/PowerSimulations.jl/src/core/formulations.jl:74.

Variables

$P_t \geq 0$: active power injected by renewable generator at time $t$ (MW)

$Q_t \geq 0$: reactive power injected by renewable generator at time $t$ (MW)

Parameters

$P^\text{max}_t$: maximum active power availability for generator at time $t$ (MW) - defined by max_active_power time series

$pf$: renewable generator power factor (see PowerSystems.get_power_factor)

Objective

Creates an objective function term based on the VariableCost Options using $G_t = P^\text{max}_t - P_t$

Constraints

\[\begin{align} & P^\text{min} \le P_t \le \eta_t P^\text{max} \\ & Q^\text{min} \le Q_t \le pf * P_t \end{align}\]